requirements.txt
python-telegram-bot
ccxt 
numpy
